# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/clubs": {
      "post": {
        "operationId": "Clubs_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClubInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Club created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or inactive category"
          },
          "409": {
            "description": "Club with this name already exists in this category"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new club",
        "tags": [
          "Clubs"
        ]
      },
      "get": {
        "operationId": "Clubs_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "example": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "search term",
              "type": "string"
            }
          },
          {
            "name": "filters",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clubs retrieved successfully"
          }
        },
        "summary": "Get all clubs with pagination",
        "tags": [
          "Clubs"
        ]
      }
    },
...
    "/api/v1/clubs/category/{categoryId}": {
      "get": {
        "operationId": "Clubs_findByCategory",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clubs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all clubs in a specific category",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/api/v1/clubs/creator/{creatorId}": {
      "get": {
        "operationId": "Clubs_findByCreator",
        "parameters": [
          {
            "name": "creatorId",
            "required": true,
            "in": "path",
            "description": "Creator User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clubs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all clubs created by a specific user",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/api/v1/clubs/{id}/stats": {
      "get": {
        "operationId": "Clubs_getStats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Club UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Club stats retrieved successfully"
          }
        },
        "summary": "Get club stats (members and projects)",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/api/v1/clubs/{id}/members": {
      "get": {
        "operationId": "Clubs_getMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Club UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Club members retrieved successfully"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get active members of a club (Leader/Admin)",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/api/v1/clubs/{id}": {
      "get": {
        "operationId": "Clubs_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Club UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Club found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Club not found"
          }
        },
        "summary": "Get a club by ID",
        "tags": [
          "Clubs"
        ]
      },
      "patch": {
        "operationId": "Clubs_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Club UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClubDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Club updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Club not found"
          },
          "409": {
            "description": "Club with this name already exists in this category"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a club (Creator/Admin only)",
        "tags": [
          "Clubs"
        ]
      },
      "delete": {
        "operationId": "Clubs_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Club UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Club deactivated successfully"
          },
          "404": {
            "description": "Club not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete a club (Creator/Admin only)",
        "tags": [
          "Clubs"
        ]
      }
    },
    "/api/v1/clubs/{id}/hard": {
      "delete": {
        "operationId": "Clubs_hardDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Club UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Club permanently deleted"
          },
          "400": {
            "description": "Cannot delete club with active members"
          },
          "404": {
            "description": "Club not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Permanently delete a club (Admin only)",
        "tags": [
          "Clubs"
        ]
      }
    },
...
    "/api/v1/courses/club/{clubId}": {
      "get": {
        "operationId": "Course_getCoursesByClub",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "description": "Club UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Club courses retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all courses for a club",
        "tags": [
          "courses"
        ]
      }
    },
...
  },
  "info": {
    "title": "CodeCircle",
    "description": "API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Root"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for external services"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Category UUID"
          },
          "name": {
            "type": "string",
            "example": "Intelligence Artificielle",
            "description": "Category name"
          },
          "slug": {
            "type": "string",
            "example": "intelligence-artificielle",
            "description": "URL-friendly slug"
          },
          "description": {
            "type": "string",
            "example": "Tout ce qui concerne l'IA et le machine learning",
            "description": "Category description"
          },
          "icon": {
            "type": "string",
            "example": "ai-icon.svg",
            "description": "Icon URL"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether the category is active"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Last update date"
          },
          "clubsCount": {
            "type": "number",
            "example": 5,
            "description": "Number of clubs in this category"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateClubInputDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "AI Masters Club",
            "description": "Club name",
            "minLength": 3,
            "maxLength": 200
          },
          "categoryId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Category UUID"
          },
          "description": {
            "type": "string",
            "example": "A club dedicated to mastering AI and machine learning",
            "description": "Club description"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://images.example.com/clubs/ai-masters.jpg",
            "description": "Club cover image URL"
          }
        },
        "required": [
          "name",
          "categoryId"
        ]
      },
      "ClubResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Club UUID"
          },
          "name": {
            "type": "string",
            "example": "AI Masters Club",
            "description": "Club name"
          },
          "categoryId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Category UUID"
          },
          "category": {
            "description": "Category details",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryResponseDto"
              }
            ]
          },
          "creatorId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Creator User UUID"
          },
          "description": {
            "type": "string",
            "example": "A club dedicated to mastering AI and machine learning",
            "description": "Club description"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://images.example.com/clubs/ai-masters.jpg",
            "description": "Club cover image URL"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether the club is active"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Last update date"
          },
          "membersCount": {
            "type": "number",
            "example": 25,
            "description": "Number of members in this club"
          },
          "projectsCount": {
            "type": "number",
            "example": 7,
            "description": "Number of projects in this club"
          }
        },
        "required": [
          "id",
          "name",
          "categoryId",
          "creatorId",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateClubDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "AI Masters Club",
            "description": "Club name",
            "minLength": 3,
            "maxLength": 200
          },
          "categoryId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Category UUID"
          },
          "description": {
            "type": "string",
            "example": "A club dedicated to mastering AI and machine learning",
            "description": "Club description"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://images.example.com/clubs/ai-masters.jpg",
            "description": "Club cover image URL"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether the club is active"
          }
        }
      },
...
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "title": {
            "type": "string",
            "example": "Advanced NestJS Development"
          },
          "description": {
            "type": "string",
            "example": "Learn advanced NestJS concepts and best practices"
          },
          "thumbnail": {
            "type": "string",
            "example": "https://example.com/course-thumbnail.jpg"
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "example": "intermediate"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "example": "published"
          },
          "duration": {
            "type": "number",
            "example": 40
          },
          "clubId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "createdBy": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "thumbnail",
          "level",
          "status",
          "duration",
          "clubId",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  },
...
}